# src/main/java/com/panpeixue/myl/model/dto/MomentCommentRequest.java
package com.panpeixue.myl.model.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class MomentCommentRequest {
    private String content;
}