# src/main/java/com/panpeixue/myl/model/pojo/MomentLike.java
package com.panpeixue.myl.model.pojo;

import java.io.Serializable;
import java.time.LocalDateTime;

public class MomentLike implements Serializable {
    private Long id;
    private Long momentId;
    private Long userId;
    private String userName;
    private LocalDateTime createTime;

    public Long getId() { return id; }
    public void setId(Long id) { this.id = id; }
    public Long getMomentId() { return momentId; }
    public void setMomentId(Long momentId) { this.momentId = momentId; }
    public Long getUserId() { return userId; }
    public void setUserId(Long userId) { this.userId = userId; }
    public String getUserName() { return userName; }
    public void setUserName(String userName) { this.userName = userName; }
    public LocalDateTime getCreateTime() { return createTime; }
    public void setCreateTime(LocalDateTime createTime) { this.createTime = createTime; }
}